+++ Views/Product/All.cshtml
@@ -15,7 +15,7 @@
                         </ol>
                     </nav>
                     <button class="navbar-toggler collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#accordionSidebar" aria-controls="accordionSidebar" aria-expanded="true" aria-label="Toggle navigation">
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="#FFF" class="bi bi-chevron-down" viewBox="0 0 16 16">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="17.5" height="17.5" fill="#FFF" class="bi bi-chevron-down" viewBox="0 0 16 16">
                             <path fill-rule="evenodd" d="M1.646 4.646a.5.5 0 0 1 .708 0L8 10.293l5.646-5.647a.5.5 0 0 1 .708.708l-6 6a.5.5 0 0 1-.708 0l-6-6a.5.5 0 0 1 0-.708z" />
                         </svg>
                     </button>
@@ -60,7 +60,6 @@
                 <div class="ItemSort row">
                     @* 數量 *@
                     <div class="col">
-                        @Model.TotalCount 項
                     </div>
                     @* Dropdown 排序 *@
                     <div class="col-auto">
