+++ Views/Shared/_Layout.cshtml
@@ -27,7 +27,7 @@
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link text-dark" href="/Product/要隱藏">所有商品</a>
+                            <a class="nav-link text-dark" href="/Product">所有商品</a>
                         </li>
                     </ul>
                     @* 新增... *@
